move throughput calc to utils
